feat: 无进有出预警

+++ app/Models/Scene.php
@@ -45,9 +45,39 @@ class Scene extends Model
 
     public static function out($visitorId, $passagewayId)
     {
-        static::whereVisitorId($visitorId)->where('passageway_id', $passagewayId)->delete();
-        Log::info('实时人员离开记录:', [
-            'scene' => '今日当前区域总人数：' . Scene::whereDate('created_at',today())->count(),
-        ]);
+        if (static::whereVisitorId($visitorId)->where('passageway_id', $passagewayId)->first()){
+            static::whereVisitorId($visitorId)->where('passageway_id', $passagewayId)->delete();
+            Log::info('实时人员离开记录:', [
+                'scene' => '今日当前区域总人数：' . Scene::whereDate('created_at',today())->count(),
+            ]);
+        }else{
+            Log::info('实时人员无进有出:', [
+                'scene' => '今日当前区域总人数：' . Scene::whereDate('created_at',today())->count(),
+            ]);
+            $visitor = Visitor::find($visitorId);
+            Warning::create([
+                'name' => $visitor->name,
+                'type' => $visitor->getType(),
+                'gender' => $visitor->gender,
+                'age' => $visitor->age,
+                'id_card' => $visitor->id_card,
+                'phone' => $visitor->phone,
+                'unit' => $visitor->unit,
+                'user_real_name' => $visitor->getUserName(),
+                'user_department' => $visitor->getUserDepartment(),
+                'reason' => $visitor->reason,
+                'access_date_from' => $visitor->access_date_from,
+                'access_date_to' => $visitor->access_date_to,
+                'ways' => $visitor->ways->pluck('name')->implode(','),
+                'access_time_from' => $visitor->access_time_from,
+                'access_time_to' => $visitor->access_time_to,
+                'limiter' => $visitor->limiter,
+                'relation' => $visitor->relation,
+                'warning_type' => '有出无进',
+                'warning_at' => now(),
+                'visitor_id' => $visitor->id
+            ]);
+        }
+
     }
 }
